social_intelligence/requirements.txt
streamlit
apify-client
python-dotenv
networkx
pyvis
requests
